ut_bt Stack Back Trace sample8.exe Menu Crash produces:
/Users/<username>/sample8_crash_log.txt

ui_app.c(2131): ui_app_exception_filter Exception: 0xC0000005(-1073741819) "Invalid access to memory location."
KERNELBASE.dll(0): UnhandledExceptionFilter ⤣
ntdll.dll(0):      memset                   ⤣
ntdll.dll(0):      _C_specific_handler      ⤣
ntdll.dll(0):      _chkstk                  ⤣
ntdll.dll(0):      RtlRaiseException        ⤣
ntdll.dll(0):      KiUserExceptionDispatcher ⤣
memcpy.asm(220):   memcpy                    ⤣
sample8.c(116):    crash                     ⤣
ui_button.c(92):   ui_button_callback        ⤣
ui_button.c(143):  ui_button_mouse           ⤣
ui_view.c(467):    ui_view_mouse             ⤣
ui_view.c(467):    ui_view_mouse             ⤣
ui_view.c(467):    ui_view_mouse             ⤣
ui_view.c(467):    ui_view_mouse             ⤣
ui_view.c(467):    ui_view_mouse             ⤣
ui_view.c(467):    ui_view_mouse             ⤣
ui_app.c(610):     ui_app_mouse              ⤣
ui_app.c(1221):    ui_app_window_proc        ⤣
USER32.dll(0):     CallWindowProcW           ⤣
USER32.dll(0):     DispatchMessageW          ⤣
ui_app.c(1469):    ui_app_message_loop       ⤣
ui_app.c(2191):    ui_app_win_main           ⤣
ui_app.c(2217):    WinMain                   ⤣
ut_bt.c(382):      ut_bt_trace_all_but_self  > Thread "RtlUserThreadStart" tid: 0x000034B8 (13496):
ntdll.dll(0):      ZwWaitForWorkViaWorkerFactory ⤣
ntdll.dll(0):      TpReleaseCleanupGroupMembers  ⤣
KERNEL32.DLL(0):   BaseThreadInitThunk           ⤣
ut_bt.c(387):      ut_bt_trace_all_but_self      < Thread "RtlUserThreadStart" tid: 0x000034B8 (13496) :
ut_bt.c(382):      ut_bt_trace_all_but_self      > Thread "RtlUserThreadStart" tid: 0x00002EF8 (12024):
ntdll.dll(0):      ZwWaitForWorkViaWorkerFactory ⤣
ntdll.dll(0):      TpReleaseCleanupGroupMembers  ⤣
KERNEL32.DLL(0):   BaseThreadInitThunk           ⤣
ut_bt.c(387):      ut_bt_trace_all_but_self      < Thread "RtlUserThreadStart" tid: 0x00002EF8 (12024) :
ut_bt.c(382):      ut_bt_trace_all_but_self      > Thread "RtlUserThreadStart" tid: 0x00002594 (9620):
ntdll.dll(0):      ZwWaitForWorkViaWorkerFactory ⤣
ntdll.dll(0):      TpReleaseCleanupGroupMembers  ⤣
KERNEL32.DLL(0):   BaseThreadInitThunk           ⤣
ut_bt.c(387):      ut_bt_trace_all_but_self      < Thread "RtlUserThreadStart" tid: 0x00002594 (9620) :
ut_bt.c(382):      ut_bt_trace_all_but_self      > Thread "ui_app.redraw" tid: 0x00005140 (20800):
ntdll.dll(0):      NtWaitForMultipleObjects      ⤣
KERNELBASE.dll(0): WaitForMultipleObjectsEx      ⤣
KERNELBASE.dll(0): WaitForMultipleObjects        ⤣
ut_threads.c(57):  ut_event_wait_any_or_timeout  ⤣
ut_threads.c(66):  ut_event_wait_any             ⤣
ui_app.c(1453):    ui_app_redraw_thread          ⤣
KERNEL32.DLL(0):   BaseThreadInitThunk           ⤣
ut_bt.c(387):      ut_bt_trace_all_but_self      < Thread "ui_app.redraw" tid: 0x00005140 (20800) :

+++ src/samples/midi.c
@@ -6,7 +6,7 @@
 
 #pragma comment(lib, "winmm")
 
-static int32_t tid; // mi is thread sensitive
+static uint64_t tid; // mi is thread sensitive
 
 ut_static_init(midi) {
     tid = ut_thread.id(); // main() thread
